% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubItemDynamicRanks.R
\name{MakeBiPartiteGraphMatrix}
\alias{MakeBiPartiteGraphMatrix}
\title{Make bi-partite graph matrix.}
\usage{
MakeBiPartiteGraphMatrix(hubItemScoresArray, hubIDColName = "HubID",
  itemIDColName = "ItemID", hubItemScoreColName = NULL,
  itemHubScoreColName = NULL, columnStochasticQ = TRUE)
}
\arguments{
\item{hubItemScoresArray}{A data frame or data table with columns...}

\item{hubIDColName}{Hub ID column name.}

\item{itemIDColName}{Item ID column name.}

\item{hubItemScoreColName}{The column name of a hub-to-item scores; if NULL contingency values are used.}

\item{itemHubScoreColName}{The column name of a item-to-hub scores; if NULL contingency values are used.}

\item{columnStochasticQ}{Should the resulting matrix be made column stochastic?}
}
\value{
A list of three elements: matrix M, hub-to-index data frame Hubs, item-to-index data frame Items.
}
\description{
Makes the adjacency matrix of a bi-partite graph connecting.
}
