% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubItemDynamicRanks.R
\name{NearestTagSets}
\alias{NearestTagSets}
\title{Nearest tag sets.}
\usage{
NearestTagSets(tagSets, tset)
}
\arguments{
\item{tagSets}{A list of lists of strings.}

\item{tset}{A list of strings.}
}
\value{
A list of lists of strings.
}
\description{
Finds the closest lists of strings to a list of strings.
}
