% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubItemDynamicRanks.R
\name{PowerMethod}
\alias{PowerMethod}
\title{Power method eigenvector computation.}
\usage{
PowerMethod(mat, bvec, alpha, maxSteps = 100, tol = 10^(-6))
}
\arguments{
\item{mat}{A sparse matrix.}

\item{bvec}{A biasing vector (a guess of the probabilities).}

\item{alpha}{The significance of \code{mat} in the iteration formula.}

\item{maxSteps}{The maximum number of iteration steps.}
}
\description{
Find the eigenvector with the largest eigenvalue using the Power method.
}
