% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubItemDynamicRanks.R
\name{RemoveZeroEntries}
\alias{RemoveZeroEntries}
\title{Remove zero entries.}
\usage{
RemoveZeroEntries(smat)
}
\arguments{
\item{smat}{A sparse matrix.}
}
\description{
Remove the zero entries from a sparse matrix.
}
