% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubItemDynamicRanks.R
\name{VectorNorm}
\alias{VectorNorm}
\title{Vector norm}
\usage{
VectorNorm(x, type = "euclidean")
}
\arguments{
\item{x}{A numeric vector.}

\item{type}{Type of norm.
One of: 'Euclidean', \code{2}, 'Infinity', \code{Inf}, 'One', code{1}.}
}
\description{
Find the Euclidean norm of a vector.
}
